deleting folder

+++ client/src/components/File.jsx
@@ -1,11 +1,12 @@
-import { useNavigate, useParams } from "react-router-dom";
+import { useLocation, useNavigate, useParams } from "react-router-dom";
 import { useState } from "react";
+
 function File(props) {
   const navigate = useNavigate();
-  // const navigateAgain = useNavigate();
   const [name, setName] = useState(props.name);
   const [showInfo, setShowInfo] = useState(false);
   const { username } = useParams();
+  const location = useLocation();
 
   function handleFolderClick(foldername) { 
     navigate(`${foldername}`);
@@ -31,18 +32,33 @@ function File(props) {
   }
 
   async function deleteFile(filename) {
+    const url = location.pathname
+    const fetchRequest = `http://localhost:4000/${url}/file/${filename}`
     try {
       const deleteFile = await fetch(
-        `http://localhost:4000/${username}/file/${filename}`,
+        fetchRequest,
         {
           method: "DELETE",
         }
       );
-      props.removeFile(filename);
+      const response = deleteFile.ok
+      response && props.removeFile(filename);
     } catch (err) {
       console.log(err);
     }
   }
+  async function deleteFolder (foldername){
+    const url=`${location.pathname.split('/')[1]}/${foldername}`;
+    console.log(url);
+    try{
+    const folder =await fetch(`http://localhost:4000/${url}`,{method:'DELETE'});
+    const response = folder.ok
+    response && props.removeFile(foldername);
+  }catch(err){
+      console.log(err)
+    }
+  }
+
   function fileInfo() {
    setShowInfo(true);
   }
@@ -66,6 +82,7 @@ function File(props) {
       ) : (<>
         <h1>Folder Name: {name}</h1>
         <button onClick={() => handleFolderClick(name)}>show folder</button>
+        <button onClick={()=>{deleteFolder(name)}}>delete folder</button>
       </>
       )}
     </div>)
